scanners/tls/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pika==1.3.2
psycopg2-binary==2.9.9
requests==2.31.0
sentinel-core @ file:///app/scanner-core
